refactor: 페이먼트 확장 설계
