+++ src/main/java/com/nhnacademy/front/shop/controller/PaymentController.java
@@ -5,12 +5,11 @@ import com.nhnacademy.front.shop.cart.client.CartClient;
 import com.nhnacademy.front.shop.order.client.OrderClient;
 import com.nhnacademy.front.shop.order.dto.OrderFormCreateRequest;
 import com.nhnacademy.front.shop.order.dto.OrderResponse;
-import com.nhnacademy.front.shop.order.service.OrderService;
 import com.nhnacademy.front.shop.payment.dto.PaymentCancelRequest;
 import com.nhnacademy.front.shop.payment.dto.PaymentConfirmRequest;
 import com.nhnacademy.front.shop.payment.dto.PaymentResponse;
 import com.nhnacademy.front.shop.payment.dto.PaymentType;
-import com.nhnacademy.front.shop.payment.dto.TossUrlProperty;
+import com.nhnacademy.front.shop.payment.dto.PaymentUrlProperty;
 import com.nhnacademy.front.shop.payment.dto.ZeroPaymentRequest;
 import com.nhnacademy.front.shop.payment.service.PaymentService;
 import com.nhnacademy.front.shop.user.client.dto.UserResponse;
@@ -21,7 +20,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -32,11 +30,9 @@ import org.springframework.web.bind.annotation.ResponseBody;
 @Controller
 @RequestMapping("/payments")
 public class PaymentController {
-    //TODO 확장성을 위해서 TossUrlProperty 수정해야함.
-    private final TossUrlProperty tossUrlProperty;
+    private final PaymentUrlProperty tossUrlProperty;
     private final PaymentService paymentService;
     private final OrderClient orderClient;
-    private final UserService userService;
     private final CartClient cartClient;
 
     @PostMapping
@@ -44,7 +40,6 @@ public class PaymentController {
     public ResponseEntity<Map<String, Object>> getPaymentPage(
             @RequestBody OrderFormCreateRequest orderFormCreateRequest
     ) {
-        UserResponse currentUser = userService.getCurrentUser();
         CommonResponse<OrderResponse> commonOrderResponse = orderClient.createOrder(orderFormCreateRequest);
         OrderResponse orderResponse = commonOrderResponse.data();
         String orderUUID = orderResponse.getOrderUUID();
@@ -55,7 +50,8 @@ public class PaymentController {
                 + "?realOrderId=" + orderResponse.getId()
                 + "&usedPointAmount=" + orderFormCreateRequest.usedPointAmount()
                 + "&discountAmount=" + orderFormCreateRequest.discountAmount()
-                + "&userId=" + currentUser.id();
+                + "&userId=" + orderFormCreateRequest.orderCreateRequest().getUserId()
+                + "&paymentType=" + orderFormCreateRequest.paymentType().name();
 
         String failUrl = tossUrlProperty.getFailURL()
                 + "?orderId=" + orderResponse.getId();
@@ -72,6 +68,7 @@ public class PaymentController {
     }
 
     //TODO  :  민감한 정보이기 때문에 레디스 사용
+    //TODO : 바로 구매 시 장바구니 안비움
     @GetMapping("/success")
     public String paymentSuccess(
             Model model,
@@ -81,7 +78,8 @@ public class PaymentController {
             @RequestParam("realOrderId") long orderId,
             @RequestParam("userId") long userId,
             @RequestParam("discountAmount") int discountAmount,
-            @RequestParam("usedPointAmount") int usedPointAmount
+            @RequestParam("usedPointAmount") int usedPointAmount,
+            @RequestParam("paymentType") String paymentType
     ) {
         PaymentConfirmRequest paymentConfirmRequest = PaymentConfirmRequest.builder()
                 .userId(userId)
@@ -91,10 +89,42 @@ public class PaymentController {
                 .discountAmount(discountAmount)
                 .usedPointAmount(usedPointAmount)
                 .amount(amount)
-                .paymentType(PaymentType.TOSS)
+                .paymentType(PaymentType.valueOf(paymentType))
                 .build();
         PaymentResponse paymentResponse = paymentService.confirmPayment(paymentConfirmRequest);
-        //TODO : 바로 구매 시 장바구니 안비움
+        cartClient.deleteCarts();
+        model.addAttribute("paymentResponse", paymentResponse);
+        return "payment/payment-success";
+    }
+
+    /**
+     * 0원 결제일 경우 처리
+     * @param model
+     * @param orderFormCreateRequest
+     * @return
+     */
+    //TODO : 바로 구매 시 장바구니 안비움
+    @PostMapping("/zero")
+    public String paymentZero(
+            Model model,
+            @RequestBody OrderFormCreateRequest orderFormCreateRequest
+    ) {
+        // 주문 생성
+        CommonResponse<OrderResponse> commonOrderResponse = orderClient.createOrder(orderFormCreateRequest);
+        OrderResponse orderResponse = commonOrderResponse.data();
+
+        // 결제 생성
+        PaymentConfirmRequest paymentConfirmRequest = PaymentConfirmRequest.builder()
+                .userId(orderResponse.getUserId())
+                .orderId(orderResponse.getId())
+                .paymentKey(null)
+                .orderUUID(orderResponse.getOrderUUID())
+                .discountAmount(orderFormCreateRequest.discountAmount())
+                .usedPointAmount(orderFormCreateRequest.usedPointAmount())
+                .amount(orderFormCreateRequest.paymentAmount())
+                .paymentType(PaymentType.valueOf(orderFormCreateRequest.paymentType().name()))
+                .build();
+        PaymentResponse paymentResponse = paymentService.confirmPayment(paymentConfirmRequest);
         cartClient.deleteCarts();
         model.addAttribute("paymentResponse", paymentResponse);
         return "payment/payment-success";
@@ -112,29 +142,5 @@ public class PaymentController {
             @RequestBody  PaymentCancelRequest cancelRequest) {
         paymentService.dealWithPaymentCancel(orderId, cancelRequest);
         return "order/order-cancel";
-    }
-
-    @PostMapping("/zero")
-    public String paymentZero(
-            Model model,
-            @RequestBody OrderFormCreateRequest orderFormCreateRequest
-    ) {
-        // 주문 생성
-        UserResponse currentUser = userService.getCurrentUser();
-        CommonResponse<OrderResponse> commonOrderResponse = orderClient.createOrder(orderFormCreateRequest);
-        OrderResponse orderResponse = commonOrderResponse.data();
-
-        // 결제 생성
-        ZeroPaymentRequest zeroPaymentRequest = ZeroPaymentRequest.builder()
-                .userId(currentUser.id())
-                .orderId(orderResponse.getId())
-                .usedPointAmount(orderFormCreateRequest.usedPointAmount())
-                .discountAmount(orderFormCreateRequest.discountAmount())
-                .amount(orderFormCreateRequest.paymentAmount())
-                .build();
-        PaymentResponse zeroPaymentResponse = paymentService.createZeroPayment(zeroPaymentRequest);
-        cartClient.deleteCarts();
-        model.addAttribute("paymentResponse", zeroPaymentResponse);
-        return "payment/payment-success";
     }
 }
