+++ src/main/java/com/nhnacademy/front/shop/guest/dto/GuestOrderFormCreateRequest.java
@@ -2,6 +2,7 @@ package com.nhnacademy.front.shop.guest.dto;
 
 import com.nhnacademy.front.shop.order.dto.OrderBookCreateRequest;
 import com.nhnacademy.front.shop.order.dto.ShipmentCreateRequest;
+import com.nhnacademy.front.shop.payment.dto.PaymentType;
 import java.util.List;
 
 public record GuestOrderFormCreateRequest(
@@ -9,6 +10,7 @@ public record GuestOrderFormCreateRequest(
     GuestOrderCreateRequest guestOrderCreateRequest,
     ShipmentCreateRequest shipmentCreateRequest,
     List<OrderBookCreateRequest> createRequestList,
-    int paymentAmount
+    int paymentAmount,
+    PaymentType paymentType
 ){
 }
