+++ src/main/java/com/nhnacademy/front/shop/payment/client/PaymentClient.java
@@ -12,9 +12,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 @FeignClient(name = "shop-service", url = "${feign.url.shop}")
 public interface PaymentClient {
-    // 결제 생성 (0원 결제 시)
-    @PostMapping("/payments")
-    CommonResponse<PaymentResponse> createZeroPayment(@RequestBody ZeroPaymentRequest request);
     // payment 결제 승인 요청
     @PostMapping("/payments/confirm")
     CommonResponse<PaymentResponse> confirmPayment(@RequestBody PaymentConfirmRequest request);
