+++ src/main/resources/templates/auth/login.html
@@ -38,7 +38,7 @@
         </form>
         <div class="mt-6 flex justify-center gap-6 text-sm text-gray-600 font-semibold">
             <a th:href="@{/register}" class="hover:text-gray-800 transition">회원가입</a>
-            <a th:href="@{/guest/login}" class="hover:text-gray-800 transition">비회원 주문조회</a>
+            <a th:href="@{/guest-order/login}" class="hover:text-gray-800 transition">비회원 주문조회</a>
         </div>
         <span class="flex items-center mt-6">
             <span class="h-px flex-1 bg-gray-400"></span>
